requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
psycopg2-binary>=2.9.0